usb: pd_engine: Fix memory leak in probe and destroy
dev_set_name() allocates memory to the provided device instance's kobject
in order to store a name. usbpd_create() returns -EPROBE_DEFER the first
time it's called, which causes it to go down the error path releasing
resources, but because device_del() never calls kobject_put() on the
kobject (it deletes it with kobject_del()), the memory allocated to the
name is never freed. Only kobject_put() frees memory allocated to the
kobject's name.

Now, replacing device_del() with device_unregister() fixes the memory leak
(because device_unregister() calls kobject_put()), but because this driver's
device doesn't have a release method, device_unregister() leaves us with
a nasty warning:
"Device 'usbpd0' does not have a release() function, it is broken and must be fixed."

Therefore, in order to fix the memory leak without angering anything,
explicitly free the memory allocated for the name with kfree_const(), which
is what kobject_put() uses to free the names it allocates.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>